Add extension point for returning different content types from API endpoints (#412)
* Add initial sketch of Codec interface.


* Introduce JSON codec.


* Expose Response type so that consuming applications (eg. Mimir) can implement their own Codecs.


* Add sketch of what supporting different codecs could look like.


* Rename fallbackCodec to defaultCodec.


* Remove defaultCodec as a field on API.


* Rename AddCodec() and clarify expected behaviour.


* Modify TestRespond to test JsonCodec directly.


* Refactor existing respond() test in preparation for content negotiation test cases.


* Add tests for content negotiation.


* Add missing documentation comments.


* Add another test case.


* Rename JsonCodec to JSONCodec.


* Fix linting issue.


* Fallback to JSON codec if no acceptable codec can be found for the Accept header.


* Move custom jsoniter code into json_codec.go.


---------
